Move schedule to topi/gpu dir
